NWBStream

Documentation for NWBStream.

NWBStream.cachefs

NWBStream.s3open

NWBStream.url2df

NWBStream.cachefs — Function

cachefs(protocol::String="http") -> CachingFileSystem

Create a caching file system object using fsspec_cached.CachingFileSystem. The protocol argument sets the underlying file system protocol to be used and is set to "http" by default.

Returns a new CachingFileSystem object that can be used for reading and writing files from the specified file system.

NWBStream.s3open — Function

s3open(s3_url::AbstractString, mode::AbstractString="rb") -> file, io

Open a file from an S3 URL.

Input arguments

s3_url::AbstractString: The S3 URL of the file.

mode::AbstractString="rb": The access mode. Default is read-only in binary mode.

Output arguments

file: A file-like NWB object

io: The io stream to the file. This should be io.close()d before exiting Julia.

Examples

s3_url = "https://visual-behavior-neuropixels-data.s3.us-west-2.amazonaws.com/visual-behavior-neuropixels/behavior_ecephys_sessions/1044385384/ecephys_session_1044385384.nwb"
data = s3open(s3_url)

NWBStream.url2df — Method

url2df(url::AbstractString) -> DataFrame

Convert the data from an online CSV file to a DataFrame.

Arguments

url::AbstractString: The URL of the CSV file.